Restructure ROS2 package into subfolders by functionality
- Move control systems components to `closed_loop/`
- Move driver controllers to `driver/`
- Move sensors to `sensors/`
- Fix any corrupted imports
